Reverted previous changes and moved the datasource packages
