+++ src/main/java/org/openapitools/datasource/Auth_DataSource.java
@@ -1,4 +1,4 @@
-package gov.dhs.nppd.humanreview.util;
+package org.openapitools.datasource;
 
 import java.util.HashMap;
